Upstream catch-up

+++ Kernel/Devices/FullDevice.cpp
@@ -5,10 +5,10 @@
  */
 
 #include <AK/Memory.h>
+#include <Kernel/API/POSIX/errno.h>
 #include <Kernel/Devices/DeviceManagement.h>
 #include <Kernel/Devices/FullDevice.h>
 #include <Kernel/Sections.h>
-#include <LibC/errno_numbers.h>
 
 namespace Kernel {
 
@@ -29,7 +29,7 @@ UNMAP_AFTER_INIT FullDevice::~FullDevice()
 {
 }
 
-bool FullDevice::can_read(const OpenFileDescription&, size_t) const
+bool FullDevice::can_read(const OpenFileDescription&, u64) const
 {
     return true;
 }
